Abstract write code into its own function.  This will make it easier to subclass functionality for both the network and USB options.
Also, make the write function try a lot harder.  The old writes would
just try to write once and then fail.  This could mean that an
incomplete packet was written to the DV3K confusing everyting.  I think
this should fix the garbage issue that was happening.  More testing is
needed on this, though.