Fix DExtra Linking
Apparently every DExtra reflector but mine wants a module ID in the
link packet otherwise it’ll NAK the connection.  This implements a
patch so that if the module doesn’t exist, we’ll pretend that it’s ‘A’.
 This may need to be propagated into the RPT1 field, but I haven’t had
a chance to test yet.  I don’t know if the reflector software checks
against linked repeaters.

This also implements a fix for a problem where if the link timer
doesn’t get shut down quickly enough, it tries to link to an empty
string.

Fixes #25