First cut at new add interface
The previous interface to add a reflector to the favorites was
cumbersome and the double click didn’t really do what was expected.
Nobody really was editing those things, so it didn’t make sense to even
keep them editable.  Double click now links/unlinks reflectors and
there is a popover for adding new reflectors.