Fix failed links
Failed links on DPlus weren’t working correctly because we were waiting
for the confirmation packet from the other side that never comes.
Moved some of the link housekeeping into the setLinkState method and
changed the network timers to directly set the state rather than trying
to send packets to the other side (which has timed out anyways).