Fix the speech synthesizer
The Vicki voice was no longer included.  We switch out to the samantha
voice instead which is much better anyways.  Samantha says “Capital”
before uppercase letters, so we lowercase strings before sending it to
her.  We also should probably strip spaces or something beforehand to
eliminate her vocalizing “space” but this is left for later on.