Changed storyboard bindings of Audio Preferences tab checkboxes from User Defaults to `AudioViewController`
Loading the Audio Preferences tab was resulting in an exception being
thrown. I couldn’t figure out exactly what was causing it, except that
it was related to the checkbox bindings to User Defaults. I addressed
the issue by exposing dynamic properties on `BTRAudioViewController`
for `courtesyTone` and `voiceAnnounceOnStatusChange` which reflect the
corresponding keys in User Defaults, then changing the checkbox
bindings away from User Defaults and to the `BTRAudioViewController`
instance, similar to the bindings for the pop-up buttons. Fixes #43

Also, I weak-ified `self` in the `BTRAudioDeviceChanged` notification
listener, to avoid that observer registration preventing the
`BTRAudioViewController` instance from deallocating.