Decrease Connect Timeout for DPlus Authenticator
The default connect timeout for TCP on MacOS X is 75s, which is way too long and causes the reflectors not to load very quickly if one of the authenticators is down for whatever reason.  Set the timeout to 8s to alleviate this problem.  8s is from the fact that OS X will send TCP SYN packets at 0s, 6s, 18s...  We don't want to make the timeout too short, otherwise we really only get a single chance for our TCP SYN to be delivered.  So, we'd like at least two of them.  That means a minimum of 6s.  We also need a chance for the SYN/ACK to get back.  Since satellite connections can sometimes have a RTT in excess of 800ms.  This means that for the SYN -> SYN/ACK RTT should be at least 800ms.  Since the TCP_CONNECTIONTIMEOUT is in seconds, we round up to 1s, and add a second for good measure.  We don't want to prematurely give up.

Also restructure the connection logic a little bit to one less redundant with the loop.